backend/requirements.txt
# Core dependencies
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0

# Database
psycopg2-binary==2.9.9
asyncpg==0.29.0
sqlalchemy[asyncio]==2.0.23
alembic==1.13.1

# Vector Database
qdrant-client==1.7.0

# LLM & Embeddings
cohere==4.46

# Web Scraping
beautifulsoup4==4.12.2
lxml==4.9.3
requests==2.31.0

# Authentication & Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6

# Utilities
python-dotenv==1.0.0
tenacity==8.2.3
aiofiles==23.2.1
markdown==3.5.1

# Development & Testing
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.2
pytest-cov==4.1.0
black==23.11.0
isort==5.12.0
flake8==6.1.0
mypy==1.7.1

# Monitoring
prometheus-client==0.19.0
structlog==23.2.0

# Rate Limiting
slowapi==0.1.9

# Performance
orjson==3.9.10